MYO-5535 (feat): moved the code
